pyproject.toml
[tool.poetry]
name = "react-langchain"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.1.0"
openai = "^1.7.2"
python-dotenv = "^1.0.0"
langchain-openai = "^0.0.2.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
